backend/requirements.txt
fastmcp==0.1.0
fastapi==0.110.0
uvicorn==0.28.0
pandas==2.2.1
faker==24.4.0
pydantic==2.6.4
google-genai==0.3.0
elevenlabs==1.0.0
